Chatbot-backend/requirements.txt
fastapi
uvicorn
pydantic
python-dotenv
langchain
langchain-huggingface
openai
langchain-openai
faiss-cpu